# Avoid extra allocation in String#from and #to