Removes unnecessary Range object allocations for a significant speed-up.

```
String#from Comparison:
  new:   3378594.0 i/s
  old:   2380129.8 i/s - 1.42x slower

String#to Comparison:
  new:   2866175.7 i/s
  old:   2304406.4 i/s - 1.24x slower
```