@kamipo 2019-05-05
Can you improve `String#to` to allocate no range object as well?